Refactor inference server and self-play runner for improved batch processing and error handling
- Updated `RustSelfPlayRunner` to support leaf batch size configuration and optimized MCTS simulation handling.
- Enhanced `InferenceServer` to ensure C-contiguous numpy arrays for safe Rust integration and improved error signaling.
- Refactored policy aggregation methods to handle cluster-based policies more effectively.
- Adjusted training and benchmarking scripts for better performance and flexibility, including increased training steps and batch sizes.
- Improved test cases to validate new functionality and ensure robustness against inference failures.

+++ native_core/src/mcts/mod.rs
@@ -370,10 +370,10 @@ impl MCTSTree {
 
             for (j, &(q, r)) in legal_moves.iter().enumerate() {
                 let ci          = first_child as usize + j;
-                let action_flat = board.window_flat_idx(q, r) as u16;
+                let action_flat = board.window_flat_idx(q, r);
                 // Look up prior from the policy vector; fall back to uniform.
-                let prior = if (action_flat as usize) < policy.len() {
-                    policy[action_flat as usize]
+                let prior = if action_flat < policy.len() {
+                    policy[action_flat]
                 } else {
                     1.0 / n_ch as f32
                 };
